Cartesian: TEMPORARY rewrite peculiarity footnote
This is incomplete and definitely needs to be redone.
Currently I am not sure whether to move the footnotes
to the main text. Doing so bloats everything,
but the general case (B arbitrary) depends on it.
It is awkward because using 'terminal curve' to refer to
the boundary of the lone non-viable region (to the right)
is technically incorrect. But if I decide to call this, say,
the 'majority terminal curve', then the captions will need
to be changed. Then for occurrences of 'majority terminal curve'
would I not put $\Phi = 0$? This is so annoying.