Remove \centredfigurecontent {extra content} argument
This commit effectively reverts 887991b6.
Having amalgamated bottom legends to the main plot,
there is no use case for this argument anymore.
The only case where a separate PDF is required for the legend
is when the legend is used across multiple subfigures.
In this case the legend is placed above the subfigures,
rather than between the main figure content and the main caption.