New branch 'we'
Basically I was in the middle of a tug-of-war between B
(who suggested avoiding 'we' lest there be an impression
that the thesis was not solely my work,
and whose thesis was written completely in the passive voice
(which completely bewildered M))
and N & M (who suggested using 'we').
This led to me using 'I', which all three of B, N, & M dislike.
After some discussion, B now agrees that I can use 'we'
in the inclusive sense (author plus reader), e.g. "We observe <X>".
So the decision is now to recast with 'we': hence this branch.