Prefer 'cases' rather than 'regimes' for radiation
Specifically, don't refer to transitions as regimes.
Thus for the Polar chapter we have three cases
(two regimes separated by one transition)
and for the Bipolar chapter we have five cases
(three regimes separated by two transitions).
This for consistency with the capillary terminology,
where the borderline case is a 'case', not a 'regime'.
Labels are changed accordingly.