Cartesian: shift cosine critical figure to next page
Having the start of the General case section
sandwiched between two figures is confusing.
Having the 'five cases:' colon pointing to the figure
rather than to the list of five cases is also confusing.
`\topfraction` is increased so as to fit the non-viable
domain figure on that page too.

More coercion needed to fit the non-viable figure on that page too.