Fix tocloft \parskip discrepancy between Debian and others
So far I have had success with reproducible builds between
  [Windows 10] TeX Live 2020
  pdfTeX 3.14159265-2.6-1.40.21 (TeX Live 2020/W32TeX)
  kpathsea version 6.3.2
and
  [Termux 0.112] texlive-full/science,science,now 20200406-4 all
  pdfTeX 3.14159265-2.6-1.40.21 (TeX Live 2020/Termux)
  kpathsea version 6.3.2
but not on
  [Debian 10] texlive-full/stable,now 2018.20190227-2 all
  pdfTeX 3.14159265-2.6-1.40.19 (TeX Live 2019/dev/Debian)
  kpathsea version 6.3.1/dev
which alone has different vertical spacing
in the Table of contents and List of figures.
After some investigation, I discovered that the discrepancy
is caused by [Debian 10] using an older version of `tocloft.sty`
which has a different treatment of `\parskip`
when the `[titles]` package option is used.
Specifically, the offending change is
  <https://github.com/wspr/herries-press/commit/6a9d617d70>,
introduced as a result of the discussion at
  <https://github.com/wspr/herries-press/issues/28>.
---
The discrepancy can be eliminated
by explicitly setting `\cftparskip` to `\parskip`.