Small: rewrite physical statement not as converse
The preceding sentence refers to replacing rough with smooth.
Therefore it is better for the physical statement
to talk of the effect of a smooth boundary, not a rough one
(which would be the converse, which is confusing).
Also fixes the line pushed over the page in b76c79a.